Reviewed by: Tim-Philipp Müller  <tim at centricular dot net> * examples/echoclient-async.c: (async_client_sin_iofunc), (async_client_in_iofunc): * examples/echoclient-gconn.c: (ob_in_iofunc): * examples/echoclient-unix.c: (normal_echoclient): * examples/echoclient.c: (main): * examples/echoserver-async.c: (async_client_iofunc): * examples/echoserver-unix.c: (normal_echoserver), (async_client_iofunc): * examples/echoserver.c: (main): * examples/hfetch.c: (hfetch): * src/conn-http.c: (gnet_conn_http_conn_connected): Fix compilation with -Werror on amd64 (mostly a simple s/guint/gsize/ for the GIOChannel API that was changed the same way in the GLib 1.2 => 2.0 transition).

+++ examples/echoclient-async.c
@@ -122,7 +122,7 @@ async_client_sin_iofunc (GIOChannel* iochannel, GIOCondition condition,
     {
       GIOError error;
       gchar buffer[1024];
-      guint bytes_read;
+      gsize bytes_read;
 
       /* Read the data into our buffer */
       error = g_io_channel_read (iochannel, buffer, 
@@ -185,7 +185,7 @@ async_client_in_iofunc (GIOChannel* iochannel, GIOCondition condition,
     {
       GIOError error;
       gchar buffer[1024];
-      guint bytes_read;
+      gsize bytes_read;
 
       /* Read the data into our buffer */
       error = gnet_io_channel_readline (iochannel, buffer, 
@@ -213,7 +213,7 @@ async_client_in_iofunc (GIOChannel* iochannel, GIOCondition condition,
 	{
 	  GIOChannel* sin;
 	  GIOError error;
-	  guint bytes_written;
+	  gsize bytes_written;
 
 	  sin = (GIOChannel*) data;
 	  error = gnet_io_channel_writen (sin, buffer, bytes_read,
